Use new selection + transform edition model
* Removed all old "Editable interface"
* use new edition interface diferent from components and entities
* use new selection model everywhere in the engine

Fixes #96

+++ src/Engine/Entity/Entity.inl
@@ -46,7 +46,7 @@ namespace Ra
 
         inline uint Entity::getNumComponents() const
         {
-            return m_components.size();
+            return uint(m_components.size());
         }
     }
